stabilize and standardize tst_qquickwidget
In tst_qquickwidget::enterLeave(), start with the cursor outside the
same as we do in QtQuick tests: move the window first, then position
the cursor to the right of it and downwards (because the window might
not necessarily go where you want it to).

Don't hard-code 5000ms timeout on qWaitForWindowExposed(): it's the
default anyway.

Task-number: QTBUG-64397
Change-Id: I67e4566c2c7f9e361a79e3a091436c3391f39786
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>