Delete textures even if one of the windows is invisible
If there is an invisible window that will not be rendered but
is constantly requested to render, updatePending is always set.
This prevented purging of already unneeded textures, and meant
endSync() was never called and m_texturesToDelete was never cleared.

Added a fix to clear the updatePending flag regardless of visibility.

Task-number: QTBUG-66116
Change-Id: Ib8bf453a3dee18d3d43342580c619664b701acae
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>