Popup: fix restoring focus for popups that did not request focus
When a popup closes, if it has active focus, it should restore focus
back to where it was regardless of whether the popup originally
requested focus. Even if a popup does not request focus on open, it
may gain focus programmatically or when a control with click-focus is
clicked, for example.

Task-number: QTBUG-66113
Change-Id: I9a7c467abe781bbef390d74898d13b9a30b2695b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>