skip filtering during post-delivery if there are no PointerHandlers
In QQuickWindowPrivate::deliverMouseEvent() we call
childMouseEventFilter on parents of items which have pointer handlers
before we give the pointer handlers one last chance to handle
un-handled events.  This last-chance delivery is a new feature in 5.10.
(Maybe letting filtering happen in that scenario is dubious anyway.
The reason it was done in 7042cfd9cb1b552c5fd753b6912439ce604eb1a0 was
to allow Flickable to steal the grab from a passively-grabbing
TapHandler.) But certainly if there aren't any pointer handlers as
children of an item, it doesn't make sense to let the item's ancestors
do filtering because in 5.9 and previous versions they would not have
gotten that chance.

Task-number: QTBUG-65651
Change-Id: I7a14a1f6cac03bf9beb4fa9ab47193df1bd773a8
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>