QmlDebug tests: Return the right value from connect(...)
If the process terminates shortly after connecting, we would return the
wrong value because the services would be NotConnected again, due to the
timeout mechanism we used to detect this. Instead, rather directly react
to the stateChanged signals and return as soon as all services are in
the correct state (or if the debug connection is dropped before, or on
timeout, which are failures).

Change-Id: I3f0c1c8519fc450627a803c76ec9b0a703104022
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>