Stabilize test
In some cases on X11, the clipboard wouldn't be empty even though
we just cleared it. The reason seems to be that it's then owned
by a different process. So simply check whether we own the selection
before checking that it's empty. This should avoid the random
failures we've been seeing in this test.

Change-Id: Id069eda5a24ad29e932aeb6a28275c20b010f90c
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>