Add Dense Material style variant for desktop
The current Material style is based on the mobile variant of the
design, which is far too large for desktop applications.

From https://material.io/guidelines/components/lists.html#lists-usage:

"When the mouse and keyboard are the primary input methods,
measurements may be condensed to accommodate denser layouts."

This patch adds a dense variant of the style where most controls like
buttons and delegates are smaller in height and use smaller font sizes.

Note that the Material design guidelines seem to distinguish between
mobile, desktop and dense measurements, where "dense" seems to be a
specialization of desktop. We cannot afford to/do not see sense in
maintaining three separate variants, so the dense variant will be the
only desktop version of the Material style.

[ChangeLog][Material] Added Dense variant of the Material style for use
on desktop platforms. Some controls are slightly smaller in height and
use smaller font sizes. The variant can be enabled by setting
QT_QUICK_CONTROLS_MATERIAL_VARIANT to Dense or setting Variant=Dense
in the qtquickcontrols.conf file.

Task-number: QTBUG-51109
Change-Id: I11846b7f6e61f7b5dcf3c146b18c220234a73ef2
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>