QQuickCheckDelegate: don't consider partially checked as checked
Same as fbcdccc for QQuickCheckBox. It was noticed missing from the
delegate counterpart while updating the "New Features in Qt 5.11"
wiki page.

[ChangeLog][Important Behavior Changes] CheckDelegate no longer considers
the partially checked state as a checked state. This fixes check state
cycling for a non-tri-state check delegate so that it goes from partially
checked to fully checked state.

Task-number: QTBUG-56295
Change-Id: I1de29178c265e027fdb91ba0b4604b110ddba06a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>