Fix QQuickTheme::font() and palette()
Respect fonts and palettes from :/qtquickcontrols2.conf, but don't
blatantly override the platform fonts and palettes with null values
if the fonts and palettes are not set in :/qtquickcontrols2.conf.

Even though fonts and palettes have technically same problem, testing
the system palette is hard with 5.11 since QQuickDefaultTheme is a
platform proxy theme and provides a custom system palette. It should
be possible to have tst_palette::systemPalette() in dev, though, where
we have removed the inheritance between QQuickTheme and QPlatformTheme.

Task-number: QTBUG-66430
Change-Id: I6dfe678ccdc7e3990320c120612cfcc68723264d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>