Move the QML_COMPILE_HASH macro definition from the command-line
If it's in the command-line, it is added to the precompiled header,
which won't be regenerated if the macro changed. Instead, let's create a
header that can be #included only in the file it needs to.

Since qmldevtools compiles the file that has this new #include, we need
to ensure it was created before the module is compiled.

Change-Id: I940917d6763842499b18fffd1513b9c64cd98387
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>