Fix tst_swipedelegate::test_eventsToLeftAndRight()
The test fails with the Fusion style and a static build on Linux.

Even if either "dx" or "dy" is defined as 0, TestCase::mouseDrag()
drags _at least_ the distance of QStyleHints::startDragDistance() in
_both_ directions. Therefore, if we start the drag from the vertical
center of the control, a horizontal drag may end below the control
and then SwipeDelegate considers the swipe canceled.

Change-Id: Ib735ebeb2184d095dfaf0a4c3bdff0bead4bdbd6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>