doc: Correct several QML property links
This update corrects several QML property links. Currently, the best
way to link to a QML property is like this:
\l {QtQuick::EventPoint::velocity} {velocity}
In other words, use the full path to the property as the link, using
'::' as the separator, and use the property name as the label. Enclose
both the link and the label in curly braces.

Change-Id: I19231403b82d41ad89096dd27787eff57c91b8ae
Reviewed-by: Martin Smith <martin.smith@qt.io>