QQuickTheme: don't inherit QPlatformTheme
Use QPlatformTheme as a fallback instead of inheriting from it. This
way, Qt Quick Controls 2 themes don't mess up the fonts and palettes
of Qt Quick Controls 1 and Qt Widgets applications.

Note: QQuickTheme::Font and QQuickTheme::Palette enums are copies of
the respective enums in QPlatformTheme, for now. This is the simplest
first step, but later on, we can have our own set of enums that cover
controls, such as Switch, that were previously entirely missing from
QPlatformTheme.

Task-number: QTBUG-51921
Change-Id: I8efe0ba2d03d65bc12b55b533ba9f2fab5320348
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>