Doc: Change QtQuickTest to QtQmlTest
The creation of module QtQmlTest remains incomplete, but this
update improves the situation a bit. In qtqmltest.qdocconf,
all instances of QtQuickTest are changed to QtQmlTest. In
qtquick.qdocconf, qtqmltest is added as a dependency. This
remains incomplete, because the module header is still called
QtQuickTest, but the number of qdoc warnings in qtdeclarative
is reduced.

Change-Id: Id545f9b5ecb2165d3f0972a059a8a4a4cfd52f2f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>