QQuickSlider: add touchDragThreshold property
Add touchDragThreshold property for configuring the threshold to initiate a 'drag',
i.e. touch move, of the handle of the slider. The mouse 'drag' won't be affected
by the property.

This property may be also used to configure the drag events stealing priorities
of nested draggable items in the future.

[ChangeLog][Slider] Added touchDragThreshold property for configuring the threshold
to initiate the touch 'drag' of the handle of the slider. The mouse 'drag' won't be
affected by the property.

Task-number: QTBUG-62784
Change-Id: Ifaeea4653b48c44efbe7e1cb09d399a91359c16a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>