QQuickWindow: skip mouse synthesis for touch events from trackpads
A trackpad is primarily a mouse-emulating device, which can also
recognize gestures, and furthermore can send raw touches iff it
is asked to.  So, every touch event from a trackpad is likely to be
followed by a mouse or gesture event (especially when the
MouseEmulation capability confirms that it is able to do that).
Therefore mouse event synthesis is redundant.

In this bug scenario, MultiPointTouchArea enables touch events within
its rectangular bounds.  A Flickable is on top.  As you flick with
two fingers on the trackpad, macOS sends touch events with two
touch points (because MPTA enabled touch events) and also a series
of QWheelEvents as it recognizes the flick gesture.  The Flickable
receives mouse events sythesized from one touch point, and reacts
as if you were dragging with one finger on a touchscreen, while it
simultaneously also reacts to the QWheelEvents.  Meanwhile the
remaining touchpoint falls through to the MPTA underneath; so the
user ends up interacting with both at the same time and making
the Flickable jump around besides.  This patch just fixes the
jumpiness in Flickable: it will no longer receive the synth-mouse
events, so it will not be dragged, and will react only to the
wheel events.  But MPTA still gets the touches.

Task-number: QTBUG-66329
Change-Id: I6f535a2c9e47bcb284eaf9ae1fdaa39f8b510af9
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>