touchDevice: don't lose bits from QTouchDevice::capabilities
It was a bug that we made a local QQuickPointerDevice::Capabilities
instance and didn't pass it to the QQuickPointerDevice constructor.
Further, we've now reserved the entire lower byte for storing
QTouchDevice::capabilities, and the upper byte for storing QtQuick-
specific capabilities, so we can copy the whole lower byte instead
of just the lower nybble.

Change-Id: Ia21b6a1503ea870a1e16cebe70a20f4e21903054
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>