Fix failing assertion when loading QML/JS on Integrity
With the QV4::Moth::BytecodeGenerator::Jump type we are relying on the
compiler doing a return value optimization. That however is not required
by the C++11 standard and the GHS compiler does indeed not do that here,
resulting in a ~Jump destructor call in the following sequence _before_
link() is called:

    Jump generateJump() { ...; return Jump(...); }
    ...
    generateJump().link();

The destructor however verifies that link() was called, which fails.

Fix this by making Jump a move-only type, which the compiler will issue
if it doesn't perform a return value optimization.

Task-number: QTBUG-66917
Change-Id: I97cc9a5d7f97d61e573ad8bc309cf48ab18eb25d
Reviewed-by: Kimmo Ollila <kimmo.ollila@qt.io>
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>